add to cart
TODO: user_id is not being read in gql side

+++ apollo/data-access/schema.ts
@@ -45,7 +45,7 @@ export const typeDefs = gql`
 
     addToCart(
       user_id: String
-      quantity: Float
+      quantity: Int
       productName: String
       price: Float
       category: String
